Ian dashboard update (#96)
* Create style for plot container

* Add data for Autuga

For the TSPlot demo

* Add code for the time series plot

* Allow accurate graphing

* Allow scrolling

* Add TSPlot to state dashboard

* Add nivo/core as dependency

* Fix line chart to allow scrolling

* Reformat data to work with component

* Move NivoTSPlot to stateDashboard

* Increase tick value to fit full screen page

* Remove scrolling

* Fix chart width

* Remove expDeathsLow and Up

* Remove padding and increase height

Increasing height to make room for legend when moving to above the graph (to preserve horizontal space)

* Update margins and move/delete legends

All to preserve horizontal space

* state-county-views

* fix-state-name

* add-padding

* sorted-selections

* created-style-file

* Update "Key Term" on landing page

Under Explanation of Key Terms, update “Coroners are elected, have limited medical training, and may be more likely to under-count Covid-19 deaths according to investigative reporting.” to “Coroners are elected and have limited medical training, whereas medical examiners are typically forensic pathologists.”

* Remove FAQ Fema Words

Remove the following text: "If you need assistance covering the cost of a COVID-19 funeral you may be eligible to apply for up to $9,000 in assistance through FEMA" from the top of the FEMA tab.

* FAQ Update Fema Words

Remove the following text: "If you need assistance covering the cost of a COVID-19 funeral you may be eligible to apply for up to $9,000 in assistance through FEMA" from the top of the FEMA tab.

* About Page - Remove Content

Remove "such as eliminating coroners" from point #2 on the Impacts Section

* Update "Who we are" on homepage

Update the "Who we are" content to the following on both the home page.  Without quotes

"This project is made possible through a collaboration between researchers at the Boston University School of Public Health and the University of Pennsylvania with support from the Boston University Center for Emerging Infectious Disease Policy & Research (CEID), BU Spark!, the Robert Wood Johnson Foundation, the W. K. Kellogg Foundation, and the National Institute on Aging. The views expressed on this website are those of the lab's investigators and do not necessarily reflect the views of the research sponsors."

* Update "Who we are" on team page

Update the "Who we are" content to the following on the team page. Without quotes

"This project is made possible through a collaboration between researchers at the Boston University School of Public Health and the University of Pennsylvania with support from the Boston University Center for Emerging Infectious Disease Policy & Research (CEID), BU Spark!, the Robert Wood Johnson Foundation, the W. K. Kellogg Foundation, and the National Institute on Aging. The views expressed on this website are those of the lab's investigators and do not necessarily reflect the views of the research sponsors."

* Update "Our team" page

Add “Alumni” tab to team page and move Sylvia Lutze, Mikas Hansen, Elif Coskun, and Steele Myrick to alumni tab.

* Update "Our team" page

Add “Alumni” tab to team page and move Sylvia Lutze, Mikas Hansen, Elif Coskun, and Steele Myrick to alumni tab.

* Remove Media Partners on Team Page

Remove all media partners tab and the people listed.

* Remove Media Partners on Team Page

Remove media partners tab and the people listed.

* Clean up removed media partners

* Clean up deleted media partners

* Clean up deleted media partners

* Clean up deleted media partners

* Remove unused import

* FAQ Page - Death Investigation

Under death investigation systems on the coroner tab, change “They do not perform any post-mortem examinations or diagnostic tests…” to “They generally do not perform post-mortem examinations or diagnostic tests...”

* Change Name of lab across website

Change "Vital Statistics Integrity Project" to "Vital Statistics Lab" everywhere it is represented. Take note that there are some instances on the team page in Dielle and Erica's bylines.

* Change Name of lab across website

Change "Vital Statistics Integrity Project" to "Vital Statistics Lab" everywhere it is represented. Take note that there are some instances on the team page in Dielle and Erica's bylines.

* Change Name of lab across website

Change "Vital Statistics Integrity Project" to "Vital Statistics Lab" everywhere it is represented. Take note that there are some instances on the team page in Dielle and Erica's bylines.

* Change Name of lab across website

Change "Vital Statistics Integrity Project" to "Vital Statistics Lab" everywhere it is represented. Take note that there are some instances on the team page in Dielle and Erica's bylines.

* Change tooltip component

* Added first parts of data wrapper stuffs

* Added datawrapper automation things

* Added smoothness to the line graph

* Updated dashboards on landing page

 - Added new dashboard style
 - Updated codebase to split landing page and dashboards
 - Added new maps and how they are shown
 - Added selector for data type on State dashboard
 - Fixed selector on County dashboard to be more flexible to use

* Applied code review changes

* Fixed file name convention issues

* Updated script to generate output file

* updated to latest version of excess mortality

* Finished templates

* WIP

* Updated national charts

* Generated one set

* Started state stuff

* Update dashboard (#93)

* Updated dashboards on landing page

 - Added new dashboard style
 - Updated codebase to split landing page and dashboards
 - Added new maps and how they are shown
 - Added selector for data type on State dashboard
 - Fixed selector on County dashboard to be more flexible to use

* Applied code review changes

* Fixed file name convention issues

* Updated national charts

* Started state stuff

Co-authored-by: Aigerim Dussikenova <aygerimabil@gmail.com>

* barchart-dashboard-interaction

* add-in-MDI-data

* Added test Cloudflare function

* Updated function

* More simple test

* another test

* Another try

* another change

* Further updates

* barchart-interactions

* install-missing-package

* Updates to state dashboard

* Update function with logging

* stringify

* restructured expected state data

* Updated script

* Updated script with note. Generated output

* Removed duplicate MDI

* Updated to fix naming of type selection

* Attempt two at fixing label

* Another fix

* Selection working

* Got all charts working. Replace iframe

* Fixed linting errors

Co-authored-by: matt-henriksen <73131815+matt-henriksen@users.noreply.github.com>
Co-authored-by: Henry Bayly <baylyh@bc.edu>
Co-authored-by: Aigerim Dussikenova <aygerimabil@gmail.com>
Co-authored-by: Aigerim Dussikenova <69381281+Aykerimm@users.noreply.github.com>